pyproject.toml
[project]
name = "example-hello-world-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "resonate-sdk>=0.6.3",
]
